requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
fastapi-cache2==0.1.9
fastapi==0.68.2
greenlet==3.0.3
h11==0.14.0
passlib==1.7.4
pendulum==3.0.0
psycopg2-binary==2.9.9
pydantic==1.10.15
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
six==1.16.0
sqlalchemy==1.4.52
starlette==0.14.2
typing-extensions==4.11.0
tzdata==2024.1
uvicorn==0.15.0
alembic==1.6.5
pre-commit==2.14.0
black==21.9b0
